+++ plugins/commands/push/command.rb
@@ -51,17 +51,19 @@ module VagrantPlugins
           if pushes.length == 1
             return pushes.first.to_sym
           else
-            raise Vagrant::Errors::PushStrategyNotProvided, pushes: pushes
+            raise Vagrant::Errors::PushStrategyNotProvided,
+              pushes: pushes.map(&:to_s)
           end
         end
 
+        name = name.to_sym
         if !pushes.include?(name)
           raise Vagrant::Errors::PushStrategyNotDefined,
-            name: name,
-            pushes: pushes
+            name: name.to_s,
+            pushes: pushes.map(&:to_s)
         end
 
-        return name.to_sym
+        return name
       end
     end
   end