+++ plugins/guests/arch/cap/configure_networks.rb
@@ -11,20 +11,14 @@ module VagrantPlugins
 
         def self.configure_networks(machine, networks)
           interfaces = Array.new
-          machine.communicate.sudo("ip -o -0 addr | grep -v LOOPBACK | awk
-                                   '{print $2}' | sed 's/://'") do |_, result|
+          machine.communicate.sudo("ip -o -0 addr | grep -v LOOPBACK | awk '{print $2}' | sed 's/://'") do |_, result|
             interfaces = result.split("\n")
           end
 
           networks.each do |network|
-            # We use :device in the template instead of
-            # eth#{network[:interface]} in order to support Predictable
-            # Network Interfaces
             network[:device] = interfaces[network[:interface]]
 
-            entry =
-                TemplateRenderer.render("guests/arch/network_#{network[:type]}",
-                                            options: network)
+            entry = TemplateRenderer.render("guests/arch/network_#{network[:type]}", options: network)
 
             temp = Tempfile.new("vagrant")
             temp.binmode
@@ -32,10 +26,8 @@ module VagrantPlugins
             temp.close
 
             machine.communicate.upload(temp.path, "/tmp/vagrant_network")
-            machine.communicate.sudo("mv /tmp/vagrant_network
-                                     /etc/netctl/#{network[:device]}")
-            machine.communicate.sudo("ip link set #{network[:device]} down &&
-                                     netctl start #{network[:device]}")
+            machine.communicate.sudo("mv /tmp/vagrant_network /etc/netctl/#{network[:device]}")
+            machine.communicate.sudo("ip link set #{network[:device]} down && netctl start #{network[:device]}")
           end
         end
       end