+++ plugins/kernel_v2/config/push.rb
@@ -21,17 +21,17 @@ module VagrantPlugins
 
         # Compile all the provider configurations
         @__defined_pushes.each do |name, tuples|
-          # Find the configuration class for this push
-          config_class = Vagrant.plugin("2").manager.push_configs[name]
-          config_class ||= Vagrant::Config::V2::DummyConfig
-
-          # Load it up
-          config = config_class.new
-
           # Capture the strategy so we can use it later. This will be used in
           # the block iteration for merging/overwriting
           strategy = name
           strategy = tuples[0][0] if tuples[0]
+
+          # Find the configuration class for this push
+          config_class = Vagrant.plugin("2").manager.push_configs[strategy]
+          config_class ||= Vagrant::Config::V2::DummyConfig
+
+          # Load it up
+          config = config_class.new
 
           begin
             tuples.each do |s, b|