+++ test/unit/plugins/provisioners/chef/omnibus_test.rb
@@ -7,14 +7,23 @@ describe VagrantPlugins::Chef::Omnibus, :focus do
 
   let(:version) { :latest }
   let(:prerelease) { false }
+  let(:download_path) { nil }
 
-  let(:build_command) { described_class.build_command(version, prerelease) }
+  let(:build_command) { described_class.build_command(version, prerelease, download_path) }
 
   context "when prerelease is given" do
     let(:prerelease) { true }
 
     it "returns the correct command" do
       expect(build_command).to eq("#{prefix} | sudo bash -s -- -p")
+    end
+  end
+
+  context "when download_path is given" do
+    let(:download_path) { '/tmp/path/to/omnibuses' }
+
+    it "returns the correct command" do
+      expect(build_command).to eq("#{prefix} | sudo bash -s -- -d \"/tmp/path/to/omnibuses\"")
     end
   end
 
@@ -34,12 +43,13 @@ describe VagrantPlugins::Chef::Omnibus, :focus do
     end
   end
 
-  context "when prerelease and version are given" do
+  context "when prerelease and version and download_path are given" do
     let(:version) { "1.2.3" }
     let(:prerelease) { true }
+    let(:download_path) { "/some/path" }
 
     it "returns the correct command" do
-      expect(build_command).to eq("#{prefix} | sudo bash -s -- -p -v \"1.2.3\"")
+      expect(build_command).to eq("#{prefix} | sudo bash -s -- -p -v \"1.2.3\" -d \"/some/path\"")
     end
   end
 end