+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+## 1.7.2 (unreleased)
+
+FEATURES:
+
+  - provisioners/salt: add support for grains [GH-4895]
+
+BUG FIXES:
+
+  - core: private boxes still referencing vagrantcloud.com will have
+      their vagrant login access token properly appended
+  - commands/push: push lookups are by user-defined name, not push
+      strategy name [GH-4975]
+  - guests/arch: fix network configuration due to poor line breaks. [GH-4964]
+  - provisioners/chef: remove Chef version check from solo.rb generation and
+      make `roles_path` populate correctly
+
+## 1.7.1 (December 11, 2014)
+
+IMPROVEMENTS:
+
+  - provisioners/ansible: Use Docker proxy if needed. [GH-4906]
+
+BUG FIXES:
+
+  - providers/docker: Add support of SSH agent forwarding. [GH-4905]
+
 ## 1.7.0 (December 9, 2014)
 
 BREAKING CHANGES:
@@ -102,6 +128,7 @@ BUG FIXES:
   - provisioners/ansible: Force `ssh` (OpenSSH) connection by default [GH-3396]
   - provisioners/ansible: Don't use or modify `~/.ssh/known_hosts` file by default,
       similarly to native vagrant commands [GH-3900]
+  - provisioners/ansible: Use intermediate Docker host when needed. [GH-4071]
   - provisioners/docker: Get GPG key over SSL. [GH-4597]
   - provisioners/docker: Search for docker binary in multiple places. [GH-4580]
   - provisioners/salt: Highstate works properly with a master. [GH-4471]